Use an adapter for gulpfile rule patching
